+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Add `isUrl` method
 - Add `isObject` method
 - Add `isEqual` method
+- Add `isEqualStrict` method
 
 ## 1.1.1 (2024-11-26)